Comeback/src/src/requirements.txt
Flask
requests
Flask-SQLAlchemy
Flask-Bcrypt
uuid
pyjwt
selenium